requirements.txt
pandas
numpy
seaborn
matplotlib
python-dotenv
pymysql
scikit-learn
-e .
